Add standard interface for routed rewrite engine (#4957)
* add RouteContextAware for ShardingSQLRewriteContextDecorator

* move ShardingSQLBuilder.getLogicAndActualTablesFromBindingTable to ShardingRule

* refactor TableToken

* refactor IndexToken

* remove LogicAndActualTablesAware

* remove generic type for SQLRouteRewriteEngine

* rename ShardingSQLBuilder to RouteSQLBuilder

* remove interface SQLRewriteEngine and SQLRouteRewriteEngine
