--- sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/sharding/rewrite/engine/ShardingSQLRewriteEngine.java
@@ -1,78 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shardingsphere.sharding.rewrite.engine;
-
-import org.apache.shardingsphere.core.rule.ShardingRule;
-import org.apache.shardingsphere.sharding.rewrite.sql.ShardingSQLBuilder;
-import org.apache.shardingsphere.underlying.common.rule.DataNode;
-import org.apache.shardingsphere.underlying.rewrite.context.SQLRewriteContext;
-import org.apache.shardingsphere.underlying.rewrite.engine.SQLRewriteResult;
-import org.apache.shardingsphere.underlying.rewrite.engine.SQLRouteRewriteEngine;
-import org.apache.shardingsphere.underlying.rewrite.parameter.builder.ParameterBuilder;
-import org.apache.shardingsphere.underlying.rewrite.parameter.builder.impl.GroupedParameterBuilder;
-import org.apache.shardingsphere.underlying.rewrite.parameter.builder.impl.StandardParameterBuilder;
-import org.apache.shardingsphere.underlying.route.context.RouteResult;
-import org.apache.shardingsphere.underlying.route.context.RouteUnit;
-
-import java.util.Collection;
-import java.util.LinkedHashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
-
-/**
- * SQL rewrite engine for sharding.
- */
-public final class ShardingSQLRewriteEngine implements SQLRouteRewriteEngine<ShardingRule> {
-    
-    @Override
-    public Map<RouteUnit, SQLRewriteResult> rewrite(final ShardingRule shardingRule, final SQLRewriteContext sqlRewriteContext, final RouteResult routeResult) {
-        Map<RouteUnit, SQLRewriteResult> result = new LinkedHashMap<>(routeResult.getRouteUnits().size(), 1);
-        for (RouteUnit each : routeResult.getRouteUnits()) {
-            result.put(each, new SQLRewriteResult(new ShardingSQLBuilder(sqlRewriteContext, shardingRule, each).toSQL(), getParameters(sqlRewriteContext.getParameterBuilder(), routeResult, each)));
-        }
-        return result;
-    }
-    
-    private List<Object> getParameters(final ParameterBuilder parameterBuilder, final RouteResult routeResult, final RouteUnit routeUnit) {
-        if (parameterBuilder instanceof StandardParameterBuilder || routeResult.getOriginalDataNodes().isEmpty() || parameterBuilder.getParameters().isEmpty()) {
-            return parameterBuilder.getParameters();
-        }
-        List<Object> result = new LinkedList<>();
-        int count = 0;
-        for (Collection<DataNode> each : routeResult.getOriginalDataNodes()) {
-            if (isInSameDataNode(each, routeUnit)) {
-                result.addAll(((GroupedParameterBuilder) parameterBuilder).getParameters(count));
-            }
-            count++;
-        }
-        return result;
-    }
-    
-    private boolean isInSameDataNode(final Collection<DataNode> dataNodes, final RouteUnit routeUnit) {
-        if (dataNodes.isEmpty()) {
-            return true;
-        }
-        for (DataNode each : dataNodes) {
-            if (routeUnit.findTableMapper(each.getDataSourceName(), each.getTableName()).isPresent()) {
-                return true;
-            }
-        }
-        return false;
-    }
-}
