+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/sharding/rewrite/token/generator/impl/ShardingInsertValuesTokenGenerator.java
@@ -19,8 +19,8 @@ package org.apache.shardingsphere.sharding.rewrite.token.generator.impl;
 
 import lombok.Setter;
 import org.apache.shardingsphere.sharding.rewrite.aware.RouteContextAware;
-import org.apache.shardingsphere.sharding.rewrite.token.pojo.impl.ShardingInsertValue;
-import org.apache.shardingsphere.sharding.rewrite.token.pojo.impl.ShardingInsertValuesToken;
+import org.apache.shardingsphere.sharding.rewrite.token.pojo.ShardingInsertValue;
+import org.apache.shardingsphere.sharding.rewrite.token.pojo.ShardingInsertValuesToken;
 import org.apache.shardingsphere.sql.parser.binder.segment.insert.values.InsertValueContext;
 import org.apache.shardingsphere.sql.parser.binder.statement.SQLStatementContext;
 import org.apache.shardingsphere.sql.parser.binder.statement.dml.InsertStatementContext;
