+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/sharding/rewrite/token/pojo/IndexToken.java
@@ -15,40 +15,63 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.rewrite.token.pojo.impl;
+package org.apache.shardingsphere.sharding.rewrite.token.pojo;
 
 import lombok.Getter;
-import org.apache.shardingsphere.sharding.rewrite.token.pojo.LogicAndActualTablesAware;
+import org.apache.shardingsphere.core.rule.ShardingRule;
+import org.apache.shardingsphere.underlying.rewrite.sql.token.pojo.RouteUnitAware;
+import org.apache.shardingsphere.sql.parser.binder.statement.SQLStatementContext;
 import org.apache.shardingsphere.sql.parser.sql.value.identifier.IdentifierValue;
 import org.apache.shardingsphere.underlying.rewrite.sql.token.pojo.SQLToken;
 import org.apache.shardingsphere.underlying.rewrite.sql.token.pojo.Substitutable;
+import org.apache.shardingsphere.underlying.route.context.RouteMapper;
+import org.apache.shardingsphere.underlying.route.context.RouteUnit;
 
+import java.util.Collection;
+import java.util.HashMap;
 import java.util.Map;
 
 /**
  * Index token.
  */
-public final class IndexToken extends SQLToken implements Substitutable, LogicAndActualTablesAware {
+public final class IndexToken extends SQLToken implements Substitutable, RouteUnitAware {
     
     @Getter
     private final int stopIndex;
     
     private final IdentifierValue identifier;
     
-    public IndexToken(final int startIndex, final int stopIndex, final IdentifierValue identifier) {
+    private final SQLStatementContext sqlStatementContext;
+    
+    private final ShardingRule shardingRule;
+    
+    public IndexToken(final int startIndex, final int stopIndex, final IdentifierValue identifier, final SQLStatementContext sqlStatementContext, final ShardingRule shardingRule) {
         super(startIndex);
         this.stopIndex = stopIndex;
         this.identifier = identifier;
+        this.sqlStatementContext = sqlStatementContext;
+        this.shardingRule = shardingRule;
     }
     
     @Override
-    public String toString(final Map<String, String> logicAndActualTables) {
+    public String toString(final RouteUnit routeUnit) {
         StringBuilder result = new StringBuilder();
         result.append(identifier.getQuoteCharacter().getStartDelimiter()).append(identifier.getValue());
+        Map<String, String> logicAndActualTables = getLogicAndActualTables(routeUnit);
         if (!logicAndActualTables.isEmpty()) {
             result.append("_").append(logicAndActualTables.values().iterator().next());
         }
         result.append(identifier.getQuoteCharacter().getEndDelimiter());
         return result.toString();
+    }
+    
+    private Map<String, String> getLogicAndActualTables(final RouteUnit routeUnit) {
+        Collection<String> tableNames = sqlStatementContext.getTablesContext().getTableNames();
+        Map<String, String> result = new HashMap<>(tableNames.size(), 1);
+        for (RouteMapper each : routeUnit.getTableMappers()) {
+            result.put(each.getLogicName().toLowerCase(), each.getActualName());
+            result.putAll(shardingRule.getLogicAndActualTablesFromBindingTable(routeUnit.getDataSourceMapper().getLogicName(), each.getLogicName(), each.getActualName(), tableNames));
+        }
+        return result;
     }
 }
